+++ src/utils/functions/crypto.ts
@@ -1,0 +1,21 @@
+import * as crypto from 'crypto';
+import { configuration } from '../../configs/configuration';
+
+const algorithm = 'aes-256-ctr';
+const secretKey = configuration().cryptoSecretKey;
+
+export function encrypt(rawData: string) {
+  const iv = crypto.randomBytes(16); // Initialisation vector
+  const cipher = crypto.createCipheriv(algorithm, secretKey, iv);
+  let encryptedData = cipher.update(rawData, 'utf-8', 'hex');
+  encryptedData += cipher.final('hex');
+  return iv.toString('hex') + encryptedData;
+}
+
+export function decrypt(encryptedData: string) {
+  const iv = Buffer.from(encryptedData.slice(0, 32), 'hex'); // Initialisation vector
+  const decipher = crypto.createDecipheriv(algorithm, secretKey, iv);
+  let decryptedData = decipher.update(encryptedData.slice(32), 'hex', 'utf-8');
+  decryptedData += decipher.final('utf-8');
+  return decryptedData;
+}